task1/requirements.txt
pyopengl
pygame
glfw
termcolor
pillow
numpy
pyrr